Move Block/Header into chain from primitives
