+++ core/beacon/src/types.rs
@@ -1,8 +1,7 @@
-use primitives::hash::{hash_struct, CryptoHash};
-use primitives::signature::{PublicKey, DEFAULT_SIGNATURE};
-use primitives::traits::{Block, Header, Signer};
-use primitives::types::{AuthorityMask, BLSSignature};
-use std::sync::Arc;
+use chain::{Block, Header};
+use primitives::hash::{CryptoHash, hash_struct};
+use primitives::signature::PublicKey;
+use primitives::types::{AuthorityMask, MultiSignature, PartialSignature};
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
 pub struct AuthorityProposal {
@@ -28,7 +27,7 @@ pub struct BeaconBlockHeaderBody {
 pub struct BeaconBlockHeader {
     pub body: BeaconBlockHeaderBody,
     pub block_hash: CryptoHash,
-    pub signature: BLSSignature,
+    pub signature: MultiSignature,
     pub authority_mask: AuthorityMask,
 }
 
@@ -40,7 +39,7 @@ pub struct BeaconBlockBody {
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
 pub struct BeaconBlock {
     pub body: BeaconBlockBody,
-    pub signature: BLSSignature,
+    pub signature: MultiSignature,
     pub authority_mask: AuthorityMask,
 }
 
@@ -67,17 +66,13 @@ impl BeaconBlock {
                     shard_block_hash
                 }
             },
-            signature: DEFAULT_SIGNATURE,
+            signature: vec![],
             authority_mask: vec![],
         }
     }
 
     pub fn genesis(shard_block_hash: CryptoHash) -> BeaconBlock {
         BeaconBlock::new(0, CryptoHash::default(), vec![], shard_block_hash)
-    }
-
-    pub fn sign(&self, signer: &Arc<Signer>) -> BLSSignature {
-        signer.sign(&self.hash())
     }
 }
 
@@ -88,13 +83,63 @@ impl Block for BeaconBlock {
         BeaconBlockHeader {
             body: self.body.header.clone(),
             block_hash: self.hash(),
-            signature: self.signature,
+            signature: self.signature.clone(),
             authority_mask: self.authority_mask.clone(),
         }
     }
+
     fn hash(&self) -> CryptoHash {
         hash_struct(&self.body)
+    }
+
+    fn add_signature(&mut self, signature: PartialSignature) {
+        self.signature.push(signature);
     }
 }
 
 pub type BeaconBlockChain = chain::BlockChain<BeaconBlock>;
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use chain::BlockChain;
+    use primitives::types::BlockId;
+    use storage::test_utils::create_memory_db;
+
+    use super::*;
+
+    #[test]
+    fn test_genesis() {
+        let storage = Arc::new(create_memory_db());
+        let genesis = BeaconBlock::new(
+            0, CryptoHash::default(), vec![], CryptoHash::default()
+        );
+        let bc = BlockChain::new(genesis.clone(), storage);
+        assert_eq!(bc.get_block(&BlockId::Hash(genesis.hash())).unwrap(), genesis);
+        assert_eq!(bc.get_block(&BlockId::Number(0)).unwrap(), genesis);
+    }
+
+    #[test]
+    fn test_restart_chain() {
+        let storage = Arc::new(create_memory_db());
+        let genesis = BeaconBlock::new(
+            0, CryptoHash::default(), vec![], CryptoHash::default()
+        );
+        let bc = BlockChain::new(genesis.clone(), storage.clone());
+        let block1 = BeaconBlock::new(
+            1, genesis.hash(), vec![], CryptoHash::default()
+        );
+        assert_eq!(bc.insert_block(block1.clone()), false);
+        let best_block = bc.best_block();
+        let best_block_header = best_block.header();
+        assert_eq!(best_block.hash(), block1.hash());
+        assert_eq!(best_block_header.hash(), block1.hash());
+        assert_eq!(best_block_header.index(), 1);
+        // Create new BlockChain that reads from the same storage.
+        let other_bc = BlockChain::new(genesis.clone(), storage.clone());
+        assert_eq!(other_bc.best_block().hash(), block1.hash());
+        assert_eq!(other_bc.best_block().header().index(), 1);
+        assert_eq!(other_bc.get_block(&BlockId::Hash(block1.hash())).unwrap(), block1);
+    }
+}
