+++ core/primitives/src/types.rs
@@ -21,8 +21,10 @@ pub type StructSignature = Signature;
 pub type MerkleHash = CryptoHash;
 /// Mask which authorities participated in multi sign.
 pub type AuthorityMask = Vec<bool>;
-/// Part of the BLS signature.
-pub type BLSSignature = Signature;
+/// Part of the signature.
+pub type PartialSignature = Signature;
+/// Whole multi signature.
+pub type MultiSignature = Vec<Signature>;
 
 pub type ReceiptId = Vec<u8>;
 pub type CallbackId = Vec<u8>;
@@ -306,7 +308,7 @@ pub struct EpochBlockHeader {
 
 #[derive(Hash, Debug)]
 pub struct SignedEpochBlockHeader {
-    pub bls_sig: BLSSignature,
+    pub bls_sig: MultiSignature,
     pub epoch_block_header: EpochBlockHeader,
 }
 
@@ -348,7 +350,7 @@ pub type TxFlowHash = u64;
 #[derive(Hash, Debug, Clone)]
 pub struct Endorsement {
     pub epoch: u64,
-    pub signature: BLSSignature,
+    pub signature: MultiSignature,
 }
 
 #[derive(Hash, Debug)]
